refactor(galgame): split plugin_config_service.py from __init__.py
Move GalgamePluginConfigService class verbatim into its own module.
__init__.py re-exports it so external import paths stay compatible.
