+++ plugin/plugins/galgame_plugin/__init__.py
@@ -97,14 +97,11 @@ from .models import (
     STORE_RAPIDOCR_AUTO_DETECT_LANG,
     STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG,
     STORE_RAPIDOCR_LANG_TYPE,
-    STORE_RAPIDOCR_OCR_VERSION,
     STORE_PUSH_NOTIFICATIONS,
     STORE_READER_MODE,
     STORE_SESSION_ID,
     json_copy,
     make_error,
-    normalize_rapidocr_ocr_version,
-    _RAPIDOCR_OCR_VERSIONS,
 )
 from .dependency_status import (
     infer_inspection_failed_dependencies,
@@ -207,133 +204,7 @@ from .plugin_capture_profile_helpers import (
 )
 
 
-class GalgamePluginConfigService:
-    def __init__(self, plugin: Any) -> None:
-        self._plugin = plugin
-
-    def persist_preferences(
-        self,
-        *,
-        bound_game_id: str,
-        mode: str,
-        push_notifications: bool,
-        advance_speed: str,
-    ) -> None:
-        self._plugin._persist.persist_preferences(
-            bound_game_id=bound_game_id,
-            mode=mode,
-            push_notifications=push_notifications,
-            advance_speed=advance_speed,
-        )
-
-    def persist_ocr_backend_selection(
-        self,
-        *,
-        backend_selection: str | None,
-        capture_backend: str | None,
-    ) -> None:
-        if backend_selection is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_OCR_BACKEND_SELECTION,
-                backend_selection,
-            )
-        if capture_backend is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_OCR_CAPTURE_BACKEND,
-                capture_backend,
-            )
-
-    def persist_rapidocr_lang(
-        self,
-        *,
-        lang_type: str | None,
-        ocr_version: str | None = None,
-        auto_detect_lang: bool | None = None,
-        auto_detect_last_lang: str | None = None,
-    ) -> None:
-        if lang_type is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_RAPIDOCR_LANG_TYPE,
-                lang_type,
-            )
-        if ocr_version is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_RAPIDOCR_OCR_VERSION,
-                ocr_version,
-            )
-        if auto_detect_lang is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_RAPIDOCR_AUTO_DETECT_LANG,
-                bool(auto_detect_lang),
-            )
-        if auto_detect_last_lang is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG,
-                auto_detect_last_lang,
-            )
-
-    def persist_reader_mode(self, *, reader_mode: str) -> None:
-        self._plugin._persist.persist_config_override(STORE_READER_MODE, reader_mode)
-
-    def persist_ocr_timing(
-        self,
-        *,
-        poll_interval_seconds: float,
-        trigger_mode: str,
-        fast_loop_enabled: bool | None = None,
-    ) -> None:
-        self._plugin._persist.persist_config_override(
-            STORE_OCR_POLL_INTERVAL_SECONDS,
-            poll_interval_seconds,
-        )
-        self._plugin._persist.persist_config_override(
-            STORE_OCR_TRIGGER_MODE,
-            trigger_mode,
-        )
-        if fast_loop_enabled is not None:
-            self._plugin._persist.persist_config_override(
-                STORE_OCR_FAST_LOOP_ENABLED,
-                bool(fast_loop_enabled),
-            )
-
-    def persist_llm_vision(
-        self,
-        *,
-        vision_enabled: bool,
-        vision_max_image_px: int,
-    ) -> None:
-        self._plugin._persist.persist_config_override(
-            STORE_LLM_VISION_ENABLED,
-            bool(vision_enabled),
-        )
-        self._plugin._persist.persist_config_override(
-            STORE_LLM_VISION_MAX_IMAGE_PX,
-            int(vision_max_image_px),
-        )
-
-    def persist_ocr_screen_templates(self, templates: list[dict[str, Any]]) -> None:
-        self._plugin._persist.persist_config_override(
-            STORE_OCR_SCREEN_TEMPLATES,
-            json_copy(templates),
-        )
-
-    def persist_runtime_state(self, payload: dict[str, Any]) -> None:
-        def _payload_int(key: str) -> int:
-            try:
-                return int(payload.get(key) or 0)
-            except (TypeError, ValueError):
-                return 0
-
-        dedupe_window = payload.get("dedupe_window")
-        last_error = payload.get("last_error")
-        self._plugin._persist.persist_runtime(
-            session_id=str(payload.get("active_session_id") or ""),
-            events_byte_offset=_payload_int("events_byte_offset"),
-            events_file_size=_payload_int("events_file_size"),
-            last_seq=_payload_int("last_seq"),
-            dedupe_window=list(dedupe_window) if isinstance(dedupe_window, list) else [],
-            last_error=dict(last_error) if isinstance(last_error, dict) else {},
-        )
+from .plugin_config_service import GalgamePluginConfigService
 
 
 @neko_plugin
@@ -2873,10 +2744,6 @@ class GalgamePlugin(NekoPluginBase):
         if value is not None:
             self._cfg.rapidocr.rapidocr_auto_detect_lang = bool(value)
 
-        value = overrides.get(STORE_RAPIDOCR_OCR_VERSION)
-        if value is not None and value in _RAPIDOCR_OCR_VERSIONS:
-            self._cfg.rapidocr.rapidocr_ocr_version = value
-
         value = overrides.get(STORE_RAPIDOCR_LANG_TYPE)
         if value is not None and value in {"ch", "japan", "korean", "en"}:
             self._cfg.rapidocr.rapidocr_lang_type = value
@@ -4688,7 +4555,7 @@ class GalgamePlugin(NekoPluginBase):
     @plugin_entry(
         id="galgame_download_rapidocr_models",
         name=tr("entries.galgame_download_rapidocr_models.name", default='下载 RapidOCR 模型'),
-        description=tr("entries.galgame_download_rapidocr_models.description", default='为当前 (lang_type, ocr_version) 选择从 ModelScope 下载缺失的 RapidOCR 模型文件到插件模型缓存目录。bundled 默认（ch+PP-OCRv4）不需要下载。'),
+        description=tr("entries.galgame_download_rapidocr_models.description", default='为当前 (lang_type, ocr_version) 选择优先从百度云下载缺失的 RapidOCR 模型文件到插件模型缓存目录，下载失败时自动回退至 ModelScope。bundled 默认（ch+PP-OCRv4）不需要下载。'),
         input_schema={
             "type": "object",
             "properties": {
@@ -5109,7 +4976,7 @@ class GalgamePlugin(NekoPluginBase):
         name=tr("entries.galgame_set_rapidocr_lang.name", default='切换 OCR 识别语言'),
         description=tr(
             "entries.galgame_set_rapidocr_lang.description",
-            default='切换 RapidOCR 文字识别语言模型和 OCR 版本；手动切换语言后关闭自动检测。',
+            default='切换 RapidOCR 文字识别语言模型；手动切换语言后关闭自动检测。',
         ),
         input_schema={
             "type": "object",
@@ -5117,10 +4984,6 @@ class GalgamePlugin(NekoPluginBase):
                 "lang_type": {
                     "type": "string",
                     "enum": ["ch", "japan", "korean", "en"],
-                },
-                "ocr_version": {
-                    "type": "string",
-                    "enum": ["PP-OCRv4", "PP-OCRv5", "v4", "v5", "4", "5"],
                 },
                 "auto_detect_lang": {
                     "type": "boolean",
@@ -5132,7 +4995,6 @@ class GalgamePlugin(NekoPluginBase):
     async def galgame_set_rapidocr_lang(
         self,
         lang_type: str | None = None,
-        ocr_version: str | None = None,
         auto_detect_lang: bool | None = None,
         **_,
     ):
@@ -5140,62 +5002,14 @@ class GalgamePlugin(NekoPluginBase):
             return Err(SdkError(self._not_configured_message()))
 
         normalized_lang = str(lang_type or "").strip().lower() or None
-        normalized_version = normalize_rapidocr_ocr_version(ocr_version) if ocr_version is not None else None
-        auto_detect_lang_provided = auto_detect_lang is not None
         if normalized_lang is not None and normalized_lang not in {"ch", "japan", "korean", "en"}:
             return Err(SdkError(f"invalid lang_type: {lang_type!r}"))
-        if ocr_version is not None and not normalized_version:
-            return Err(SdkError(f"invalid ocr_version: {ocr_version!r}"))
-        if normalized_version is not None and normalized_version == self._cfg.rapidocr_ocr_version:
-            normalized_version = None
-        if (
-            normalized_lang is not None
-            and normalized_lang == self._cfg.rapidocr_lang_type
-            and not self._cfg.rapidocr_auto_detect_lang
-        ):
-            normalized_lang = None
-        if (
-            normalized_lang is None
-            and auto_detect_lang_provided
-            and bool(auto_detect_lang) == bool(self._cfg.rapidocr_auto_detect_lang)
-            and normalized_version is None
-        ):
-            return Ok({
-                "lang_type": self._cfg.rapidocr_lang_type,
-                "ocr_version": self._cfg.rapidocr_ocr_version,
-                "auto_detect_lang": self._cfg.rapidocr_auto_detect_lang,
-                "summary": (
-                    f"RapidOCR lang={self._cfg.rapidocr_lang_type} "
-                    f"ocr_version={self._cfg.rapidocr_ocr_version} "
-                    f"auto_detect={'on' if self._cfg.rapidocr_auto_detect_lang else 'off'}"
-                ),
-                "skipped": True,
-                "already_applied": True,
-                "skip_reason": "already_applied",
-            })
-        if normalized_lang is None and auto_detect_lang is None and normalized_version is None:
-            if lang_type is None and ocr_version is None and not auto_detect_lang_provided:
-                return Err(SdkError("lang_type, ocr_version or auto_detect_lang is required"))
-            return Ok({
-                "lang_type": self._cfg.rapidocr_lang_type,
-                "ocr_version": self._cfg.rapidocr_ocr_version,
-                "auto_detect_lang": self._cfg.rapidocr_auto_detect_lang,
-                "summary": (
-                    f"RapidOCR lang={self._cfg.rapidocr_lang_type} "
-                    f"ocr_version={self._cfg.rapidocr_ocr_version} "
-                    f"auto_detect={'on' if self._cfg.rapidocr_auto_detect_lang else 'off'}"
-                ),
-                "skipped": True,
-                "already_applied": True,
-                "skip_reason": "already_applied",
-            })
+        if normalized_lang is None and auto_detect_lang is None:
+            return Err(SdkError("lang_type or auto_detect_lang is required"))
 
         old_lang = self._cfg.rapidocr_lang_type
-        old_version = self._cfg.rapidocr_ocr_version
         old_auto = self._cfg.rapidocr_auto_detect_lang
         old_last = self._cfg.rapidocr_auto_detect_last_lang
-        if normalized_version is not None:
-            self._cfg.rapidocr_ocr_version = normalized_version
         if normalized_lang is not None:
             self._cfg.rapidocr_lang_type = normalized_lang
             self._cfg.rapidocr_auto_detect_last_lang = normalized_lang
@@ -5208,7 +5022,6 @@ class GalgamePlugin(NekoPluginBase):
                 self._ocr_reader_manager.update_config(self._cfg)
             except Exception as exc:
                 self._cfg.rapidocr_lang_type = old_lang
-                self._cfg.rapidocr_ocr_version = old_version
                 self._cfg.rapidocr_auto_detect_lang = old_auto
                 self._cfg.rapidocr_auto_detect_last_lang = old_last
                 return Err(SdkError(f"apply rapidocr lang failed: {exc}"))
@@ -5221,7 +5034,6 @@ class GalgamePlugin(NekoPluginBase):
         try:
             self._config_service.persist_rapidocr_lang(
                 lang_type=normalized_lang,
-                ocr_version=normalized_version,
                 auto_detect_lang=(
                     bool(auto_detect_lang)
                     if normalized_lang is None and auto_detect_lang is not None
@@ -5231,7 +5043,6 @@ class GalgamePlugin(NekoPluginBase):
             )
         except Exception as exc:
             self._cfg.rapidocr_lang_type = old_lang
-            self._cfg.rapidocr_ocr_version = old_version
             self._cfg.rapidocr_auto_detect_lang = old_auto
             self._cfg.rapidocr_auto_detect_last_lang = old_last
             if self._ocr_reader_manager is not None:
@@ -5250,11 +5061,9 @@ class GalgamePlugin(NekoPluginBase):
         self._start_background_bridge_poll()
         return Ok({
             "lang_type": self._cfg.rapidocr_lang_type,
-            "ocr_version": self._cfg.rapidocr_ocr_version,
             "auto_detect_lang": self._cfg.rapidocr_auto_detect_lang,
             "summary": (
                 f"RapidOCR lang={self._cfg.rapidocr_lang_type} "
-                f"ocr_version={self._cfg.rapidocr_ocr_version} "
                 f"auto_detect={'on' if self._cfg.rapidocr_auto_detect_lang else 'off'}"
             ),
         })
